// src/types/database.d.ts
export interface o_script {
 'content'?: string | null;
 'createTime'?: number | null;
 'errorReason'?: string | null;
 'extractState'?: number | null;
 'id'?: number;
 'name'?: string | null;
 'projectId'?: number | null;
}